Support gmail checkerplus notifications
Action icons are nice to have so we've bundled them. We need to detect the translated action text, which isn't the best way to go, but works for me/english speakers in the short term.

Chrome notifications also elide and don't wrap the list of emails when it shows multiple emails. Since emails tend to be very long, only 1 summary is visible usually. So we need to not wrap the notification body based on the action icon text.

Also provide another test for the checkerplus stuff and action buttons.